Added abstract IndexDocument builder; Changed abstractions order;
